+++ Products/CMFPlone/exportimport/tests/testPropertiesTool.py
@@ -7,7 +7,6 @@ _PROPERTYSHEET_XML = """\
 <?xml version="1.0"?>
 <object name="site_properties" meta_type="Plone Property Sheet">
  <property name="title">Site wide properties</property>
- <property name="allowAnonymousViewAbout" type="boolean">True</property>
  <property name="displayPublicationDateInByline"
     type="boolean">True</property>
 </object>
@@ -18,7 +17,6 @@ _PROPERTIESTOOL_XML = """\
 <object name="portal_properties" meta_type="Plone Properties Tool">
  <object name="site_properties" meta_type="Plone Property Sheet">
   <property name="title">Site wide properties</property>
-  <property name="allowAnonymousViewAbout" type="boolean">True</property>
   <property name="displayPublicationDateInByline"
      type="boolean">True</property>
  </object>
@@ -35,8 +33,8 @@ class PropertySheetXMLAdapterTests(BodyAdapterTestCase):
 
     def _populate(self, obj):
         obj.manage_changeProperties(title='Site wide properties')
-        obj.manage_addProperty('allowAnonymousViewAbout', True, 'boolean')
-        obj.manage_addProperty('displayPublicationDateInByline', True, 'boolean')
+        obj.manage_addProperty(
+            'displayPublicationDateInByline', True, 'boolean')
 
     def setUp(self):
         from Products.CMFPlone.interfaces import ISimpleItemWithProperties
@@ -62,10 +60,8 @@ class PropertiesToolXMLAdapterTests(BodyAdapterTestCase):
                        SimpleItemWithProperties('site_properties'))
         obj.site_properties.manage_changeProperties(
             title='Site wide properties')
-        obj.site_properties.manage_addProperty('allowAnonymousViewAbout',
-                                               True, 'boolean')
-        obj.site_properties.manage_addProperty('displayPublicationDateInByline',
-                                               True, 'boolean')
+        obj.site_properties.manage_addProperty(
+            'displayPublicationDateInByline', True, 'boolean')
 
     def setUp(self):
         from Products.CMFPlone.exportimport.propertiestool \
@@ -74,7 +70,8 @@ class PropertiesToolXMLAdapterTests(BodyAdapterTestCase):
         from Products.CMFPlone.interfaces import ISimpleItemWithProperties
         from Products.GenericSetup.interfaces import ISetupEnviron
         from Products.GenericSetup.interfaces import IBody
-        provideAdapter(self._getTargetClass(), (IPropertiesTool, ISetupEnviron), IBody)
+        provideAdapter(
+            self._getTargetClass(), (IPropertiesTool, ISetupEnviron), IBody)
         provideAdapter(
             SimpleItemWithPropertiesXMLAdapter,
             (ISimpleItemWithProperties, ISetupEnviron), IBody)